python/csp_solver/requirements.txt
selenium==3.14.0
geopandas==0.4.0
matplotlib>=1.5.3
pandas==0.21.0
PyHamcrest==1.9.0
descartes==1.1.0
future>=0.16.0
